Implemented getObject for ValueAccess, sharing underlying implementation with PgResultSet.
